TreeLSTMs and Clinical Data merge
Trained the combined dataset on:
- Logistical Regression classifier
- Cox and DeepSurv models
- Random Forest classifier

Main takeaways:
Best results seen with Random Forest due to feature importance visualizations.
Logistical Regression is a poor model to use as a baseline for the TreeLSTM dataset as it is small and therefore overfits. I will use a standard Bayesian Probabilistic Model instead as the baseline to compare our more sophisticated Random Forest model.